+++ Infrastructure/Documents/Excel/Month.cs
@@ -1,0 +1,17 @@
+﻿namespace Infrastructure.Documents.Excel;
+
+public enum Month
+{
+    Januar = 1,
+    Februar,
+    März,
+    April,
+    Mai,
+    Juni,
+    Juli,
+    August,
+    September,
+    Oktober,
+    November,
+    Dezember
+}
